Move RequestQuery into domain.requests.query
